requirements.txt
fastApi
uvicorn
requests
python-dotenv
SQLAlchemy
Psycopg2-binary
alembic
apscheduler
pydantic
apscheduler
